scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "scripts for preprocessing fortune cookies"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "langchain>=0.3.9",
    "langchain-openai>=0.2.10",
    "loguru>=0.7.2",
    "pytrends>=4.9.2",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "langchain-community>=0.3.8",
    "pydantic>=2.10.2",
    "dashscope>=1.20.14",
    "requests-cache>=1.2.1",
]
